+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 - Add `random` method
 - Add `findById` method
-- support colloquial color names for color identity param
+- Support colloquial color names for color identity param
+- `search`
+  - Support `=`, `>`, `<`, `>=`, `<=` operators for `coloridentity`
 
 # 0.2.0